filter ShowNamespaces by custom PlanerStrategy
